get the last release from github releases for the current branch (#1)
Imagine running this tool from github actions. With this PR, the tool is
able to get the current branch name and the current github repo from
github actions environment. Then use that info to get the last github
release made on that branch.

This info is considered the single source of truth for the last
successful deployment.

# Testing 

No automated tests written, as there is no logic to test yet. I am
thinking of mocking the git and github modules so the only code that
needs tests is `deploy.ts` and there's no logic in there yet.

So only testing done thus far is running the deploy script and test it
against some repos to see if it gives back a valid git tag and commit
sha.