make this script into a github action (#4)
Create GitHub Action so we can execute the script. 

1. Create a UI for the Action logging messages when it runs. 
2. Create `action.yml` 
3. Run Deno code in docker container. 

# Testing 

Install action in another repo: `uses:
levibostian/next-deployment-tool@create-action` and run it in different
scenarios.

I was successfully able to create a `1.0.0` release. However when trying
to create `1.1.0`, I found a bug where the script was not able to find
the old `1.0.0` release when it should have. That bug fix will come in
another PR.