fix: find last github release for branch (#5)
# Problem 

Code has a bug where it cannot get the last github release that was made
on a branch.

In last deployment attempt, I noticed that `taget_commitish` for github
release fetch results returned a commit hash instead of branch name as
expected.

# Solution 

The solution to fix this was much more complex then I thought!

- Get list of latest 100 or whatever commits for the current branch. 
- Get the list of latest 100 github releases. 
- Get list of latest 100 git tags. 
- Map the releases and tags so we get a list of github releases with
it's matching commit hashes.
- Iterate the list of commits and see if there is a match with the
commit hash. if there is, great, found it! If not, get the next latest
release-tag and loop again.
- All while having good paging code so we iterate all commit pages and
we do not waste network calls.

# Testing 

- Automated tests have been written for the logic. Then, after we merge
this code, the code will run and we will see if it fixes it.