refactor: improve log messages highlighting the steps that the tool takes (#8)
## Related GitHub Issues
<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem
<!-- A clear description of the problem that this pull request is
solving. -->

When viewing the logs when the action runs, the logs are not very
helpful. Ideally, you could glance at them and understand the steps that
are taken and what logs are important and what are not as important.

## Solution
<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Using colors and emojis to make each step stand out. 

Before: 
![CleanShot 2024-08-20 at 07 56
23](https://github.com/user-attachments/assets/4aa6e5e5-a9f5-4208-ba56-5ae54185565a)
![CleanShot 2024-08-20 at 07 57
44](https://github.com/user-attachments/assets/fa566053-2703-4189-818f-c2efd8cb481b)

After: 
![CleanShot 2024-08-21 at 07 13
26](https://github.com/user-attachments/assets/d9c670b2-688d-403c-8fa2-cf6deb8cae2e)


## Testing
<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->
- [ ] Added automated tests. 
- [X] Manually tested. If you check this box, provide instructions for
others to test, too.

* Ran via dry-run to confirm that I see color in the logs. 
* Ran this branch on a testing repo I have to see what the logs look
like in there.

## Notes for reviewers 
<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->