refactor: run CLI outside docker container (#9)
## Related GitHub Issues
<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem
<!-- A clear description of the problem that this pull request is
solving. -->

The next feature to add to the tool is being able to run scripts
provided by the user to execute to deploy their code. In order for us to
be able to run scripts, we must run the CLI outside of a Docker
container. Running scripts inside a Docker container will simply not
work or be too complex to work as we intend.

## Solution
<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Since this tool is written in Deno, I evaluated a few options...


1. (What I chose) Change the action to be a composite action that
installs Deno runtime and then executes the deploy CLI script via Deno
CLI.

pros: easy to maintain, simple to understand, get the deno features like
runtime permissions, pretty fast compared to maybe running binaries, run
CLI the way deno intends which gives us all deno features if we need
them, if there is a bug with tool the stacktrace is accurate
cons: does this make the tool more complex by requiring that we do
runtime permissions?

<details>
<summary>List of other options not chosen</summary>
<br>
2. Compile a binary of this CLI tool and then have the action run the
binary

pros: for the end developer, it's just running a CLI. easy for them. no
runtime permissions to worry about
cons: harder to maintain since I need to create and release binaries.
when bugs occur for users, is it harder for them to report issues since
the stacktrace might be weird?

4. Generate [npm module of this tool](https://deno.com/learn/modules),
upload to npm, and then simply run `npx new-deployment-tool` in the
github action.
pros: pretty simple to maintain. node installed on github actions by
default. running npx is simple.
cons: more work during deployments but minimal. dont get runtime
permissions. maybe the tool is buggy and it will limit the deno code I
could run

5. Upload this tool to jsr > in github action, use `npx jsr install
new-deployment-tool` > `npx new-deployment-tool`
pros: just upload deno code to jsr and it does the work for me.
cons: runtime permissions work or break? not sure if jsr is meant to
work well for CLI tools but packages.
</details>

## Testing
<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->
- [ ] Added automated tests. 
- [X] Manually tested. If you check this box, provide instructions for
others to test, too.

Ran this branch on a test repository. It ran a deployment as expected. 

## Notes for reviewers 
<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->