feat: pass context to deploy commands (#12)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

Let's say that your deployment script is going to modify the
`package.json` file with the new semantic version for this new release
and then push the code to npmjs.com. Well, that means your script needs
to know the new version that we are trying to release. This PR adds that
functionality of sending contextual data to the deployment scripts.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

This tool's design is around flexibility. Part of that is allowing you
to run scripts in whatever language you want. From that, here are some
of the design decisions of this PR...

- Pass input data into each of the scripts via an environment variable
containing a JSON string. Each language out there can access environment
variables and can parse JSON. Therefore, it's the idea behind how we can
communicate with scripts from any language you choose. (coming in
another PR).

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [x] Manually tested. If you check this box, provide instructions for
others to test, too.

In a private repo I use for testing, I created a python script that
parses the input JSON string and would output the new version to assert
that it received the contextual data as expected:

```py
import os
import json

# Test you can get the input 

input_data = json.loads(os.environ.get("INPUT"))
next_version_name = input_data.get("nextVersionName")
print(next_version_name)
```

Personally, I think this is pretty cool. Only 2 lines of code to add to
your script to be able to receive input data in a language beyond
Typescript and Deno that this tool is written in!

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#12 :point\_left:
      - \#14
        - \#15
          - \#16