refactor: communicate with deploy commands with a file (#14)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

We are about to implement a feature where the deploy commands can send
data *back* to the deployment tool instead of just having the tool send
data to the commands. I have been trying to figure out how to do this
and the only way I could come up with that was still simple to do was to
use a file that each the tool and the commands read and write.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

The commands are given an environment variable that contains a path to
the data file. That file contains the JSON string that they can read to
get all the contextual data they need. The idea is that commands will
then also write to this file to send data back to the tool.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [x] Manually tested. If you check this box, provide instructions for
others to test, too.

Ran python code very similar to the automated tests where we print the
JSON file from the environment variable path. When running inside of a
real github action, it was hard to get the runtime permissions correct.
Manually tested was the only way to validate that.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#14 :point\_left:
      - \#15
        - \#16