feat: commands can send data back to the tool (#15)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

When a command runs, it may need to communicate back to the tool.
Because we are allowing commands to be written in any language you want,
we need to have some common way to do this communication.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

The tool communicates with deploy commands by the commands reading a
JSON file. We re-use this concept where the command can write data to
this same file and the tool will then read from that file.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [x] Manually tested. If you check this box, provide instructions for
others to test, too.

Ran branch on a private github repo to test this works. Ran a python
script `json.dump({"filesToCommit": ["foo.txt"]},
open(os.getenv("DATA_FILE_PATH"), "w"))` and verified that the tool was
able to parse it and return it from exec.run successfully.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#15 :point\_left:
      - \#16