feat: git commit files modified during deploy commands (#16)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

For some type of deployments, you may need to edit metadata files and
require those file changes to be in the deployment. Example: Swift
Package Manager projects since the git tag is the deployment.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

The deployment script when it runs gets to send the tool back data. As
part of this, it can tell the tool what files should be added,
committed, and pushed. The tool performs all of the git operations for
you.

All of the scripts are executed and only 1 commit is created. The
committer is the github actions bot at this time. We can allow you to
customize this later.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.

The only automated test missing is "Given a git commit was created,
expect that the github release created points to that new git commit,
not the old one". A refactor of the codebase needs to be done to do this
test and I consider that out of scope of this already large PR.

- [x] Manually tested. If you check this box, provide instructions for
others to test, too.

The automated test missing above, has been manually tested in a test
repo.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

### Other solutions I tried

At first, I thought about just making the deployment scripts be
responsible for doing all of the git operations. Since this project
tries to lean towards having less features then more that feel optional.
However, there are many reasons why this solution is not ideal.

1. There are a lot of git operations to run. And because we are running
the tool in CI environment where the committer email/name is not
populated by default, it's easy for the user to mess up and break their
deployments.
2. The user may have multiple deployment scripts to run where each need
to commit files. It would be preferred if each script can independently
say what files they need committed, but still only 1 git commit is
created by running this tool.
3. This tool needs to keep track of the latest commit at all times. We
need a reliable solution where the tool is able to keep track of it,
even when new commits are made.

<!-- branch-stack -->

- `main`
  - \#11
    - \#16 :point\_left: