chore: run deno fmt (#19)
## Related GitHub Issues
<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem
<!-- A clear description of the problem that this pull request is
solving. -->

Not sure why, but formatter didnt run for previous PRs. I installed git
hooks again, we will see if that fixes it.

## Solution
<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

## Testing
<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->
- [ ] Added automated tests. 
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers 
<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->