test: write integration tests of all steps done during deployment (#18)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

I am wanting to test that all of the steps of the deployment process
work correctly together. These are those tests!

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Write tests for `deploy.ts`, which is the top-level file in the codebase
that orchestrates the deployment process.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#18 :point\_left: