refactor: safer git operations by using a dedicated branch (#21)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

Users can run scripts to deploy their software. Some deployment
processes require committing file changes.

One of the missions of this project is *calm deployments*. Running git
commands on someone's git repo has risks and responsibilities. If any of
those commands fails, that could lead to the git repo being in a weird
state that the user must fix themselves in a manual way. These scenarios
are not calm. It would be ideal if we refactor the tool to perform these
git operations in a less risky way.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Run all git operations on a git branch that this tool controls. By
committing all changes to a particular branch...
...if the deployment is successful, the user can automatically or
manually merge the git commit we create into their `main` branch.
...if the deployment is not successful, the worst case is that a new git
branch got created that just sits there.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [X] Added automated tests.
- [X] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#21 :point\_left: