fix: improve the logs to explain the deployment process better (#22)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

Part of the vision of this project is great logs that walk you through
the entire deployment process. Logs that are hopefully so good, you
don't need to read any of the documentation of the project.

After reviewing the logs to this point, they could be better.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

I evaluated all of the logs that are shown to the user and edited them
to be more human readable to someone who is perhaps brand new to a
project and this is their first deployment. The logs should be much more
user friendly.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.

I refactored the test suite so we now perform snapshot testing on the
logs. Allowing us to more easily evaluate the logs and improve them over
time.

- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#22 :point\_left: