fix: deleting commits for previous deployments if branch not merged (#23)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

With the current code base, this scenario can happen:

- Deployment 1.0 happens today. During that deployment, a new git commit
was made on the git branch that we control.
- The new git commit has not yet been merged into `main`, as the user
has not yet merged it.
- Deployment 1.1 happens later today. During this deployment, a new git
commit will be created. The tool up until this point uses only 1
hard-coded branch name for all new commits made. And we force push to
cleaned failed deployments. This means that we will delete the branch
from 1.0 which includes the git commit that the git tag 1.0 points to.
This will break the git tag 1.0.

Now, because the 1.0 tag has not yet been merged, this will probably
mean that when 1.1 is attempted to run, it will actually try to deploy
1.0 again. We do strongly suggest to users that before running 1.0 a
second time, they do a check to make sure that they have not already ran
that deployment. But, we still *really* do not want to ever break a git
tag. So even though this scenario may not happen, if it ever does, it's
bad. So just to be safe, make sure that we do not break git tags, ever.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

For each deployment version, create a unique tag. A compromise between
not making too many branches in the user's git repo while also trying to
not break git tags.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11
    - \#23 :point\_left: