feat: run provided scripts for deployment (#11)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

This PR is what allows you to actually perform your deployment. After
the tool has determined that a new release should occur, the tool will
run all of your scripts that you have provided for deployment.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

This tool's design is around flexibility. Part of that is allowing you
to run scripts in whatever language you want. From that, here are some
of the design decisions of this PR...

- You provide bash commands to the tool which allows you to run any
language you want. Example: `python deploy.py`
- Pass input data into each of the scripts via an environment variable
containing a JSON string. Each language out there can access environment
variables and can parse JSON. Therefore, it's the idea behind how we can
communicate with scripts from any language you choose. (coming in
another PR).
- If any of the scripts fail, the tool exists early.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [x] Manually tested. If you check this box, provide instructions for
others to test, too.

In a private repo I use for testing, I used this workflow to run the
tool:

```yml
      - name: Deploy
        uses: levibostian/new-deployment-tool@send-input-to-plugin
        with:
          github_token: ${{ secrets.GITHUB_TOKEN }}
          version: send-input-to-plugin
          deploy_commands: |
            pip install -r requirements.txt
            python test.py
```

To test that the tool can run all of the commands as if I were to run
those commands myself in the workflow.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#11 :point\_left: