docs: write getting started guide (#24)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

The tool is now at the point where it has enough functionality that it
can be used in some projects. The problem is, there are no docs telling
you how to use the tool.

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Write a getting started guide in the README explaining how to install
the tool and write deployment scripts for your project.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [ ] Added automated tests.
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

<!-- branch-stack -->

- `main`
  - \#24 :point\_left: