feat: deploy pre-release versions (#25)
## Related GitHub Issues

<!-- Link to any related GitHub issues that this pull request addresses
or closes. -->

## Problem

<!-- A clear description of the problem that this pull request is
solving. -->

While developing a project, it's nice to be able to make deployments
when the code is not ready for production. Example: Release an early
beta or alpha of some code. This PR allows you to do that!

## Solution

<!-- Describe the approach you took to solve the problem and the changes
made in this pull request. -->

Similar to semantic-release's strategy, you can designate specific
branches to indicate the type of release you want to make. So if I want
to make a beta release, I can create a branch `beta` and push code to it
and it will generate a new beta release.

This tool currently requires you use conventional commits and semantic
versioning of your code. This PR updates the functionality of our use of
semantic versioning to generate versions that include prerelease
versions. I suggest viewing the unit tests for *determine next release
step* to see all the improvements made.

## Testing

<!-- Choose one of the below options for how you tested the code change.
Include any specific setup or instructions for testing. -->

- [x] Added automated tests.
- [ ] Manually tested. If you check this box, provide instructions for
others to test, too.

## Notes for reviewers

<!-- If there is any additional information you would like to share with
the person reviewing this pull request, please provide it here. -->

Semantic versioning gets quite complex once you add prerelease versions
into the mix. Because of that, I decided to not reinvent the wheel and
begin using a 3rd party semantic version module.

The problem is, the module's increment feature did not match the
behavior that I was looking for so I needed to write some logic to make
it work. This logic is based heavily on semantic-release's logic.

<!-- branch-stack -->

- `main`
  - \#25 :point\_left: